Update Renderer to support drawing in a separate thread and various updates/fixes
* Imagine: Add Gfx::RendererCommands to hold the renderer's draw state
* Imagine: Add Gfx::RendererTask to control the renderer's draw context and thread
* Imagine: Add Gfx::SyncFence to keep resources in sync with draw thread
* Imagine: Add View::prepareDraw() for loading image resources before drawing
* Imagine: Remove Gfx::RenderTarget, obsoleted by setting render to texture in RendererCommands
* Imagine: Use a dedicated thread to manage the main OpenGL context
* Imagine: Optionally use a separate thread and context when drawing with RendererTask
* Imagine: Add support for OpenGL share contexts
* Imagine: Add Gfx::Text::makeGlyphs()
* Imagine: Make Gfx::Texture movable
* Imagine: Make Base::Pipe more robust and movable
* Imagine: Make Semaphore movable
* Imagine: Add EventLoop::stop()
* Imagine: Fix GlibEventLoop with multiple threads using a thread local GMainContext
* Imagine: Add SimpleFrameTimer for use with systems with no OS-level frame time events
* Imagine: Use SimpleFrameTimer on Android versions below 4.1 and remove obsolete eventfd frame timer
* Imagine: Use a thread local variable to get & store the JNIEnv pointer
* Imagine: Use .aar output file instead of .jar and update Gradle scripts for Android
* Imagine: Increase Android x86 minimum SDK level to 16 for NDK 18
* Imagine: Don't manage the framebuffer object in EAGLView since it must be created on the draw thread
* Imagine: Properly disconnect the CAEAGLLayer when deleting the renderbuffer
* Imagine: Update docs to list Xcode 10 requirement due to C++17 features
* Imagine: Remove deprecated X11 GLX code and always use EGL
* Imagine: Update bundled X11 libraries for Pandora build
* Imagine: Add colored logger output for stderr
* Imagine: Add size template parameter to DelegateFunc as DelegateFunc2
* Imagine: Add DelegateFuncSet to easier support multiple delegates with priorities
* EmuFramework: Don't run emulation in draw handler since it now runs on a separate thread
* EmuFramework: Remove dither video option and control it automatically based on the active video effect
* EmuFramework: Add Renderer multithreading option
* EmuFramework: Fix build without CONFIG_VCONTROLS_GAMEPAD
* EmuFramework: Add option to control multithreading in renderer

+++ imagine/include/imagine/util/audio/PcmFormat.hh
@@ -55,27 +55,27 @@ public:
 		return frames * sample.toBytes() * channels;
 	}
 
-	float framesToMSecs(uint frames) const
+	double framesToMSecs(uint frames) const
 	{
-		return ((float)frames / rate) * 1000.;
+		return ((double)frames / rate) * 1000.;
 	}
 
-	float framesToUSecs(uint frames) const
+	double framesToUSecs(uint frames) const
 	{
-		return ((float)frames / rate) * 1000000.;
+		return ((double)frames / rate) * 1000000.;
 	}
 
-	uint mSecsToFrames(float mSecs) const
+	uint mSecsToFrames(double mSecs) const
 	{
-		return std::ceil((mSecs / 1000.f) * rate);
+		return std::ceil((mSecs / 1000.) * rate);
 	}
 
-	uint uSecsToFrames(float uSecs) const
+	uint uSecsToFrames(double uSecs) const
 	{
-		return std::ceil((uSecs / 1000000.f) * rate);
+		return std::ceil((uSecs / 1000000.) * rate);
 	}
 
-	uint uSecsToBytes(float uSecs) const
+	uint uSecsToBytes(double uSecs) const
 	{
 		return framesToBytes(uSecsToFrames(uSecs));
 	}
@@ -90,9 +90,9 @@ public:
 		return (rate * sample.toBytes() * channels) * secs;
 	}
 
-	float bytesToSecs(uint bytes) const
+	double bytesToSecs(uint bytes) const
 	{
-		return (float)bytes / (rate * sample.toBytes() * channels);
+		return (double)bytes / (rate * sample.toBytes() * channels);
 	}
 };
